requirements.txt
django==5.1.1
djangorestframework==3.15.2
django-environ==0.11.2
PyMySQL==1.1.1
mypy==1.11.2
flake8==7.1.1
black==24.8.0
drf-writable-nested==0.7.0
djangorestframework-camel-case==1.4.2